requirements.txt
Flask==0.12.2
kafka-python==1.4.2

